requirements.txt

# packages used in the app:
python-dotenv
flask


# packages used for testing:
pytest